Split up script
