+++ game/src/main/kotlin/content/social/trade/exchange/GrandExchangeOffers.kts
@@ -2,42 +2,34 @@ package content.social.trade.exchange
 
 import com.github.michaelbull.logging.InlineLogger
 import content.entity.npc.shop.stock.ItemInfo
-import content.entity.player.bank.bank
 import content.entity.player.bank.isNote
 import content.entity.player.bank.noted
-import content.entity.player.dialogue.type.intEntry
 import content.entity.player.inv.item.tradeable
 import content.entity.player.modal.Tab
 import content.entity.player.modal.tab
 import world.gregs.voidps.engine.client.message
 import world.gregs.voidps.engine.client.sendScript
-import world.gregs.voidps.engine.client.ui.InterfaceOption
-import world.gregs.voidps.engine.client.ui.close
 import world.gregs.voidps.engine.client.ui.dialogue.continueItemDialogue
 import world.gregs.voidps.engine.client.ui.event.interfaceClose
 import world.gregs.voidps.engine.client.ui.event.interfaceOpen
 import world.gregs.voidps.engine.client.ui.interfaceOption
 import world.gregs.voidps.engine.client.ui.open
-import world.gregs.voidps.engine.data.Settings
 import world.gregs.voidps.engine.data.definition.ItemDefinitions
 import world.gregs.voidps.engine.entity.character.player.Player
-import world.gregs.voidps.engine.entity.character.player.chat.inventoryFull
-import world.gregs.voidps.engine.entity.character.player.chat.notEnough
-import world.gregs.voidps.engine.entity.character.player.name
 import world.gregs.voidps.engine.entity.item.Item
 import world.gregs.voidps.engine.entity.playerSpawn
 import world.gregs.voidps.engine.inject
-import world.gregs.voidps.engine.inv.clear
 import world.gregs.voidps.engine.inv.inventory
 import world.gregs.voidps.engine.inv.sendInventory
-import world.gregs.voidps.engine.inv.transact.TransactionError
-import world.gregs.voidps.engine.inv.transact.operation.AddItemLimit.addToLimit
-import world.gregs.voidps.engine.inv.transact.operation.RemoveItem.remove
-import world.gregs.voidps.engine.inv.transact.operation.RemoveItemLimit.removeToLimit
 import kotlin.math.ceil
 
 val exchange: GrandExchange by inject()
 val itemDefinitions: ItemDefinitions by inject()
+val logger = InlineLogger()
+
+/*
+    Offers
+ */
 
 playerSpawn { player ->
     exchange.login(player)
@@ -52,44 +44,8 @@ interfaceOpen("grand_exchange") { player ->
     exchange.refresh(player)
 }
 
-interfaceOption("Collect*", "collect_slot_*", "grand_exchange") {
-    val index = component.removePrefix("collect_slot_").toInt()
-    val box: Int = player["grand_exchange_box"] ?: return@interfaceOption
-    val id: Int = player["grand_exchange_offer_${box}"] ?: return@interfaceOption
-    val collectionBox = player.inventories.inventory("collection_box_${box}")
-    val item = collectionBox[index]
-    var noted = item
-    if (option == "Collect_notes") {
-        noted = item.noted ?: item
-    }
-    player.inventory.transaction {
-        val txn = link(collectionBox)
-        val added = addToLimit(noted.id, item.amount)
-        txn.remove(item.id, added)
-    }
-    when (player.inventory.transaction.error) {
-        is TransactionError.Full -> player.inventoryFull()
-        TransactionError.None -> if (collectionBox.isEmpty()) {
-            val offer = exchange.offers.offer(id)
-            if (offer != null && offer.state.cancelled) {
-                exchange.offers.remove(id)
-                player.clear("grand_exchange_offer_${box}")
-                clear(player)
-            }
-            exchange.refresh(player, box)
-        }
-        else -> logger.warn { "Issue collecting items from grand exchange ${player.inventory.transaction.error} ${player.name} $item $index" }
-    }
-}
-
-interfaceOption("Abort Offer", "offer_abort", "grand_exchange") {
-    val slot: Int = player["grand_exchange_box"] ?: return@interfaceOption
-    abort(player, slot)
-}
-
-interfaceOption("Abort Offer", "view_offer_*", "grand_exchange") {
-    val slot = component.removePrefix("view_offer_").toInt()
-    abort(player, slot)
+interfaceClose("grand_exchange") {
+    GrandExchange.clear(it)
 }
 
 interfaceOption("Make Offer", "view_offer_*", "grand_exchange") {
@@ -99,6 +55,11 @@ interfaceOption("Make Offer", "view_offer_*", "grand_exchange") {
     player["grand_exchange_box"] = slot
     selectItem(player, offer.item)
 }
+
+
+/*
+    Buy Offer
+ */
 
 interfaceOption("Make Buy Offer", "buy_offer_*", "grand_exchange") {
     val slot = component.removePrefix("buy_offer_").toInt()
@@ -117,6 +78,17 @@ fun openItemSearch(player: Player) {
     player.sendScript("item_dialogue_reset", "Grand Exchange Item Search")
 }
 
+continueItemDialogue { player ->
+    selectItem(player, item)
+    player["grand_exchange_price"] = player["grand_exchange_market_price", 0]
+    ItemInfo.showInfo(player, Item(item))
+}
+
+
+/*
+    Sell Offer
+ */
+
 interfaceOption("Make Sell Offer", "sell_offer_*", "grand_exchange") {
     val slot = component.removePrefix("sell_offer_").toInt()
     player["grand_exchange_box"] = slot
@@ -133,8 +105,6 @@ interfaceOpen("stock_side") { player ->
     player.sendScript("grand_exchange_hide_all")
 }
 
-val logger = InlineLogger()
-
 interfaceOption("Offer", "items", "stock_side") {
     val item = if (item.isNote) item.noted else item
     if (item == null) {
@@ -150,171 +120,6 @@ interfaceOption("Offer", "items", "stock_side") {
     player["grand_exchange_price"] = player["grand_exchange_market_price", 0]
 }
 
-interfaceOption("Back", "back", "grand_exchange") {
-    clear(player)
-}
-
-interfaceOption("Confirm Offer", "confirm", "grand_exchange") {
-    val slot: Int = player["grand_exchange_box"] ?: return@interfaceOption
-    val itemId: String = player["grand_exchange_item"] ?: return@interfaceOption
-    val amount: Int = player["grand_exchange_quantity"] ?: return@interfaceOption
-    val price: Int = player["grand_exchange_price"] ?: return@interfaceOption
-    val id = when (player["grand_exchange_page", "offers"]) {
-        "buy" -> {
-            val total = price * amount
-            var fromBank = false
-            player.inventory.transaction {
-                var removed = removeToLimit("coins", total)
-                if (removed < total && Settings["grandExchange.useBankCoins", false]) {
-                    val txn = link(player.bank)
-                    removed += txn.removeToLimit("coins", total - removed)
-                    fromBank = true
-                }
-                if (removed < total) {
-                    error = TransactionError.Deficient(total - removed)
-                }
-            }
-            when (player.inventory.transaction.error) {
-                TransactionError.None -> {
-                    if (fromBank) {
-                        player.message("Payment has been taken from your bank.")
-                    }
-                    exchange.buy(player, Item(itemId, amount), price)
-                }
-                is TransactionError.Deficient -> {
-                    println(player.inventory.transaction.error)
-                    player.notEnough("coins")
-                    return@interfaceOption
-                }
-                else -> return@interfaceOption
-            }
-        }
-        "sell" -> {
-            var removed = 0
-            player.inventory.transaction {
-                removed += removeToLimit(itemId, amount)
-                if (removed < amount) {
-                    val noted = Item(itemId, amount).noted?.id ?: itemId
-                    removed += removeToLimit(noted, amount - removed)
-                }
-
-                if (removed < amount) {
-                    error = TransactionError.Deficient(amount - removed)
-                }
-            }
-            when (player.inventory.transaction.error) {
-                TransactionError.None -> exchange.sell(player, Item(itemId, amount), price)
-                else -> {
-                    logger.warn { "Error removing GE items ${player.name} ${player.inventory.transaction.error} $slot $itemId $amount $price" }
-                    return@interfaceOption
-                }
-            }
-        }
-        else -> return@interfaceOption
-    }
-    player.inventories.inventory("collection_box_${slot}").clear()
-    player["grand_exchange_offer_${slot}"] = id
-    exchange.refresh(player, slot)
-    clear(player)
-}
-
-interfaceClose("grand_exchange") {
-    clear(it)
-}
-
-fun clear(player: Player) {
-    player["grand_exchange_box"] = -1
-    player["grand_exchange_page"] = "offers"
-    player.sendScript("item_dialogue_close")
-    player.close("stock_side")
-    player.close("item_info")
-    player.clear("grand_exchange_item")
-    player.clear("grand_exchange_item_id")
-    player.clear("grand_exchange_price")
-    player.clear("grand_exchange_market_price")
-    player.clear("grand_exchange_range_min")
-    player.clear("grand_exchange_range_max")
-    player.clear("grand_exchange_quantity")
-}
-
-interfaceOption("Add *", "add_*", "grand_exchange") {
-    when (player["grand_exchange_page", "offers"]) {
-        "sell" -> {
-            val total = totalItems()
-            player["grand_exchange_quantity"] = when (component) {
-                "add_1" -> 1
-                "add_10" -> 10
-                "add_100" -> 100
-                "add_all" -> total
-                else -> return@interfaceOption
-            }.coerceAtMost(total)
-        }
-        "buy" -> when (component) {
-            "add_1" -> player.inc("grand_exchange_quantity", 1)
-            "add_10" -> player.inc("grand_exchange_quantity", 10)
-            "add_100" -> player.inc("grand_exchange_quantity", 100)
-            "add_all" -> player.inc("grand_exchange_quantity", 1000)
-            else -> return@interfaceOption
-        }
-    }
-}
-
-interfaceOption("Edit Quantity", "add_x", "grand_exchange") {
-    player["grand_exchange_quantity"] = when (player["grand_exchange_page", "offers"]) {
-        "sell" -> intEntry("Enter the amount you wish to sell:").coerceAtMost(totalItems())
-        "buy" -> intEntry("Enter the amount you wish to purchase:")
-        else -> return@interfaceOption
-    }
-}
-
-interfaceOption("Increase Quantity", "increase_quantity", "grand_exchange") {
-    if (player["grand_exchange_quantity", 0] < Int.MAX_VALUE - 1) {
-        player["grand_exchange_quantity"] = (player["grand_exchange_quantity", 0] + 1).coerceAtMost(totalItems())
-    }
-}
-
-interfaceOption("Decrease Quantity", "decrease_quantity", "grand_exchange") {
-    if (player.dec("grand_exchange_quantity", 1) < 0) {
-        player["grand_exchange_quantity"] = 0
-    }
-}
-
-interfaceOption("Decrease Price", "decrease_price", "grand_exchange") {
-    player["grand_exchange_price"] = (player["grand_exchange_price", 0] - 1).coerceAtLeast(player["grand_exchange_range_min", 0])
-}
-
-interfaceOption("Increase Price", "increase_price", "grand_exchange") {
-    player["grand_exchange_price"] = (player["grand_exchange_price", 0] + 1).coerceAtMost(player["grand_exchange_range_max", 0])
-}
-
-interfaceOption("Offer Market Price", "offer_market", "grand_exchange") {
-    player["grand_exchange_price"] = player["grand_exchange_market_price", 0]
-}
-
-interfaceOption("Edit Price", "offer_x", "grand_exchange") {
-    player["grand_exchange_price"] = when (player["grand_exchange_page", "offers"]) {
-        "sell" -> intEntry("Enter the price you wish to sell for:")
-        "buy" -> intEntry("Enter the price you wish to buy for:")
-        else -> return@interfaceOption
-    }.coerceIn(player["grand_exchange_range_min", 0], player["grand_exchange_range_max", 0])
-}
-
-// TODO: Toggle to replace min/max with +/- 5%
-
-interfaceOption("Offer Minimum Price", "offer_min", "grand_exchange") {
-    player["grand_exchange_price"] = player["grand_exchange_range_min", 0]
-}
-
-interfaceOption("Offer Maximum Price", "offer_max", "grand_exchange") {
-    player["grand_exchange_price"] = player["grand_exchange_range_max", 0]
-}
-
-continueItemDialogue { player ->
-    selectItem(player, item)
-    player["grand_exchange_price"] = player["grand_exchange_market_price", 0]
-    ItemInfo.showInfo(player, Item(item))
-}
-
 fun selectItem(player: Player, item: String) {
     val definition = itemDefinitions.get(item)
     player["grand_exchange_item"] = item
@@ -324,22 +129,4 @@ fun selectItem(player: Player, item: String) {
     player["grand_exchange_market_price"] = price
     player["grand_exchange_range_min"] = ceil(price * 0.95).toInt()
     player["grand_exchange_range_max"] = ceil(price * 1.05).toInt()
-}
-
-fun abort(player: Player, slot: Int) {
-    val id: Int = player["grand_exchange_offer_${slot}"] ?: return
-    exchange.cancel(id)
-    // https://youtu.be/3ussM7P1j00?si=IHR8ZXl2kN0bjIfx&t=398
-    player.message("Abort request acknowledged. Please be aware that your offer may have already been completed.")
-}
-
-fun InterfaceOption.totalItems(): Int {
-    val item = Item(player["grand_exchange_item", ""])
-    val noted = item.noted
-    var total = 0
-    if (noted != null) {
-        total += player.inventory.count(noted.id)
-    }
-    total += player.inventory.count(item.id)
-    return total
 }
